Do not wait for the first byte of an empty body
If the body of a message is empty, no body will ever be sent, so waiting
for one to arrive is wrong.  The symptom is that the GUI agent appears
to run one message behind, but *only* for empty messages, which is
extremely confusing.  Fix this by immediately returning the message if
it is an empty one that is known to this library, or otherwise
immediately trying to read the next header.