Agent: switch from a handler trait to an event enum
A trait object is essentially a table of callbacks, and is a _very_ poor
choice for this application.  Not only does it cause problems with the
borrow checker, it also makes for much more complicated code, and makes
bindings for other languages (such as OCaml, for Mirage OS) difficult to
implement.

Instead, use an enum that holds all of the possible events.  That allows
for much simpler client code, and avoids the need to thread state (such
as the ‘Client’ or ‘Agent’ objects) through callbacks.  It also plays
nicely with language bindings, as calling Rust from another language is
often much faster than calling that language from Rust.